Fixed api bugs

+++ src/main/java/com/encooked/services/AuthorityService.java
@@ -27,7 +27,7 @@ import org.springframework.stereotype.Service;
  */
 @Service
 public class AuthorityService {
-    
+
     @Autowired
     UserService userService;
 
@@ -40,64 +40,78 @@ public class AuthorityService {
     @Autowired
     GrantedPriviledgeEntityRepository grantedPriviledgeEntityRepository;
 
-    public List<String> getAllRoles() {
+    public List<String> getAllRoles(boolean indicateSystem) {
         return roleEntityRepository.findAll()
-                .stream().map(p -> p.getName())
+                .stream().map(p -> p.getName() + (indicateSystem ? p.isSystem() ? "*" : "" : ""))
                 .collect(Collectors.toList());
     }
-    
-     public List<String> getAllRolesByPriviledge(String value) throws RecordNotFoundException {
+
+    public List<String> getAllRoles() {
+        return getAllRoles(false);
+    }
+
+    public List<String> getAllRolesByPriviledge(String value, boolean indicateSystem) throws RecordNotFoundException {
         return grantedPriviledgeEntityRepository.findByValue(value)
-                .orElseThrow(() -> new RecordNotFoundException("Priviledge " + value + " does not exist"))
-                .getRoles().stream().map(p -> p.getName())
+                .stream()
+                .map(p -> p.getRole().getName() + (indicateSystem ? p.getRole().isSystem() ? "*" : "" : ""))
+                .distinct()
                 .collect(Collectors.toList());
+    }
+
+    public List<String> getAllRolesByPriviledge(String value) throws RecordNotFoundException {
+        return getAllRolesByPriviledge(value, false);
     }
 
     public List<String> getAllAuthorities() {
         List<String> grantedAuthorities = new ArrayList<>();
         roleEntityRepository.findAll().forEach(role -> {
             role.getPriviledges().forEach(priv -> {
-                String a = role.getName() +"."+priv.getValue();
-                if(priv.isRead()){
-                    a = a.toUpperCase()+"_READ";
-                    grantedAuthorities.add(a);
+                String a = role.getName() + "." + priv.getValue();
+                if (priv.isRead()) {
+                    grantedAuthorities.add(a.toUpperCase() + "_READ");
                 }
-                
-                if(priv.isWrite()){
-                   a = a.toUpperCase()+"_READ";
-                   grantedAuthorities.add(a);
+
+                if (priv.isWrite()) {
+                    grantedAuthorities.add(a.toUpperCase() + "_WRITE");
                 }
             });
         });
         return grantedAuthorities;
     }
 
-    
+    public List<String> getAllPermission(boolean indicateSystem) {
+        return priviledgeEntityRepository.findAll()
+                .stream().map(p -> p.getValue() + (indicateSystem ? p.isSystem() ? "*" : "" : ""))
+                .collect(Collectors.toList());
+    }
 
     public List<String> getAllPermission() {
-        return priviledgeEntityRepository.findAll()
-                .stream().map(p -> p.getValue())
+        return getAllPermission(false);
+    }
+
+    public List<String> getAllPermissionsByRole(String role, boolean indicateSystem) throws RecordNotFoundException {
+        return roleEntityRepository.findByName(role)
+                .orElseThrow(() -> new RecordNotFoundException("Role " + role + " does not exist"))
+                .getPriviledges().stream().map(p -> p.getValue() + (indicateSystem ? p.isSystem() ? "*" : "" : ""))
                 .collect(Collectors.toList());
     }
 
     public List<String> getAllPermissionsByRole(String role) throws RecordNotFoundException {
-        return roleEntityRepository.findByName(role)
-                .orElseThrow(() -> new RecordNotFoundException("Role " + role + " does not exist"))
-                .getPriviledges().stream().map(p -> p.getValue())
-                .collect(Collectors.toList());
+        return getAllPermissionsByRole(role, false);
     }
 
-    public RoleEntity createRole(String name) throws RecordNotFoundException {
-        roleEntityRepository.findByName(name)
-                .orElseThrow(() -> new RecordNotFoundException(String.format("Role \"%s\" already exist", name)));
-
+    public RoleEntity createRole(String name) throws Exception {
+        if (roleEntityRepository.findByName(name).isPresent()) {
+            throw new Exception(String.format("Role '%s\' already exist", name));
+        }
         RoleEntity role = new RoleEntity();
         role.setName(name);
         return roleEntityRepository.save(role);
     }
 
     private void validateAuthority(String value) throws Exception {
-        if (value.split(".").length != 2) {
+        value = value.trim();
+        if (value.split("\\.").length != 2) {
             throw new Exception("Invalid Authority.Valid format: ROLE.PERMISSION_ACCESS Example: \"ADMIN.USER_READ\"");
         }
 
@@ -105,12 +119,12 @@ public class AuthorityService {
             throw new Exception("Invalid Authority.Valid format: ROLE.PERMISSION_ACCESS Example: \"ADMIN.USER_READ\"");
         }
 
-        Pattern p = Pattern.compile("[^A-Z0-9\\-\\_]", Pattern.CASE_INSENSITIVE);
-        if (p.matcher(value).find()) {
+        Pattern p = Pattern.compile("^[a-zA-Z0-9._-]*$", Pattern.CASE_INSENSITIVE);
+        if (!p.matcher(value).matches()) {
             throw new Exception("Invalid Authority: Allowed characters; a to z, A to Z, 0 to 9,\"_\" \"_\" and \"-\"  Examples: \"ADMIN.USER_READ\", \"ADMIN.WEB-SOCKET_READ\"");
         }
 
-        if (!value.endsWith("_READ") || !value.endsWith("_WRITE")) {
+        if (!value.endsWith("_READ") && !value.endsWith("_WRITE")) {
             throw new Exception("Invalid Authority: Must end with \"_READ\" or \"_WRITE\"  Examples: \"ADMIN.USER_READ\", \"ADMIN.WEB-SOCKET_WRITE\"");
         }
     }
@@ -119,7 +133,7 @@ public class AuthorityService {
         String value = authority.getAuthority().toUpperCase();
         validateAuthority(value);
 
-        String[] split = authority.getAuthority().split(".");
+        String[] split = authority.getAuthority().split("\\.");
 
         String role = split[0];
         String priviledge_access = split[1];
@@ -135,15 +149,18 @@ public class AuthorityService {
         priviledgeEntityRepository.findByValue(priviledge)
                 .orElseThrow(() -> new RecordNotFoundException((String.format("Priviledge \"%s\" does not exist", priviledge))));
 
-        GrantedPriviledgeEntity grantedPriviledgeEntity = grantedPriviledgeEntityRepository
-                .findByValue(priviledge).orElse(new GrantedPriviledgeEntity(value));
+        GrantedPriviledgeEntity grantedPriviledgeEntity = grantedPriviledgeEntityRepository.findByValue(priviledge)
+                .stream()
+                .filter(p -> p.getRole().getName().equals(role))
+                .findFirst().orElseGet(() -> {
+                    return grantedPriviledgeEntityRepository.save(new GrantedPriviledgeEntity(priviledge));
+                });
 
         grantedPriviledgeEntity.setRead(access.equals("READ") ? grant : grantedPriviledgeEntity.isRead());
         grantedPriviledgeEntity.setWrite(access.equals("WRITE") ? grant : grantedPriviledgeEntity.isWrite());
 
-        grantedPriviledgeEntity.getRoles().add(roleEntity);
+        grantedPriviledgeEntity.setRole(roleEntity);
         roleEntity.getPriviledges().add(grantedPriviledgeEntity);
-
         roleEntityRepository.save(roleEntity);
 
         return authority;
@@ -158,22 +175,31 @@ public class AuthorityService {
     }
 
     public int deleteRole(String name) throws Exception {
-        if(!userService.getAllUsersByRole(name).isEmpty()){
+        if (!userService.getAllUsersByRole(name).isEmpty()) {
             throw new Exception("You cannot delete roles that are already in use");
+        }
+        RoleEntity role = roleEntityRepository.findByName(name).orElseThrow(() -> new Exception("Role does not exist"));
+        if (role.isSystem()) {
+            throw new Exception("You cannot delete a system role");
         }
         return roleEntityRepository.deleteByName(name).size();
     }
-    
+
     public int deletePriviledge(String value) throws Exception {
-        if(!getAllRolesByPriviledge(value).isEmpty()){
+        if (!getAllRolesByPriviledge(value).isEmpty()) {
             throw new Exception("You cannot delete priviledges that are already in use");
+        }
+        PriviledgeEntity priviledge = priviledgeEntityRepository.findByValue(value).orElseThrow(() -> new Exception("Priviledge does not exist"));
+        if (priviledge.isSystem()) {
+            throw new Exception("You cannot delete a system priviledge");
         }
         return priviledgeEntityRepository.deleteByValue(value).size();
     }
 
-    public PriviledgeEntity createPriviledge(String value) throws RecordNotFoundException {
-        priviledgeEntityRepository.findByValue(value)
-                .ifPresent(p -> new RecordExistsException((String.format("Priviledge \"%s\" already exist", p))));
+    public PriviledgeEntity createPriviledge(String value) throws Exception {
+        if (priviledgeEntityRepository.findByValue(value).isPresent()) {
+            throw new Exception(String.format("Priviledge '%s\' already exist", value));
+        }
 
         PriviledgeEntity priviledgeEntity = new PriviledgeEntity();
         priviledgeEntity.setValue(value);
